fix: sidebar scroll otel hub
