+++ app/(opentelemetry-hub-routes)/guides/loading.tsx
@@ -1,0 +1,1 @@
+export { default } from '@/components/GlobalLoader/GlobalLoader'
