+++ app/(opentelemetry-hub-routes)/guides/page/[page]/page.tsx
