--- app/guides/loading.tsx
@@ -1,1 +1,0 @@
-export { default } from "@/components/GlobalLoader/GlobalLoader"
