+++ layouts/open-telemetry-hub/OpenTelemetrySidebarClient.tsx
@@ -58,7 +58,7 @@ export default function OpenTelemetrySidebarClient({
   useEffect(() => {
     setTriggerContainer(document.getElementById(mobileTriggerId))
     setOverlayContainer(document.getElementById(mobileOverlayId))
-  }, [mobileOverlayId, mobileTriggerId])
+  }, [mobileOverlayId, mobileTriggerId, normalizedRoute])
 
   useEffect(() => {
     if (typeof window === 'undefined') return
